+++ LinguagemDeProgramacao/data-cat-jar/src/main/java/data/cat/modal/Maquina.java
@@ -1,7 +1,3 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package data.cat.modal;
 
 /**
@@ -10,16 +6,16 @@ package data.cat.modal;
  */
 public class Maquina {
 
-    private Long idMaquina;
+    private Integer idMaquina;
     private String indentPessoal;
     private String hostName;
     private String grupo;
-    private Long fkEmpresa;
+    private Integer fkEmpresa;
 
     public Maquina() {
     }
 
-    public Maquina(Long idMaquina, String indentPessoal, String hostName, String grupo, Long empresa) {
+    public Maquina(Integer idMaquina, String indentPessoal, String hostName, String grupo, Integer empresa) {
         this.idMaquina = idMaquina;
         this.indentPessoal = indentPessoal;
         this.hostName = hostName;
@@ -27,11 +23,11 @@ public class Maquina {
         this.fkEmpresa = empresa;
     }
 
-    public Long getIdMaquina() {
+    public Integer getIdMaquina() {
         return idMaquina;
     }
 
-    public void setIdMaquina(Long idMaquina) {
+    public void setIdMaquina(Integer idMaquina) {
         this.idMaquina = idMaquina;
     }
 
@@ -59,11 +55,11 @@ public class Maquina {
         this.grupo = grupo;
     }
 
-    public Long getEmpresa() {
+    public Integer getEmpresa() {
         return fkEmpresa;
     }
 
-    public void setEmpresa(Long empresa) {
+    public void setEmpresa(Integer empresa) {
         this.fkEmpresa = empresa;
     }
